requirements.txt


# NLP e pré-processamento
nltk==3.8.1

# Embeddings e IA
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# LangChain (versão modular + compatível)
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.35

# Clusterização
hdbscan==0.8.33

asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.6
joblib==1.3.2
keybert==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==23.2
pillow==11.2.1
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
pandas==2.2.2
nltk==3.8.1
langchain==0.1.14
sentence-transformers==2.2.2
# faiss-cpu==1.7.4